requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.2
flask-cors==4.0.0
langchain==0.0.348
pypdf==3.17.4
huggingface-hub==0.14.1
sentence-transformers==2.2.2
torch==1.13.1+cpu
transformers==4.30.2
python-docx==0.8.11
markdown==3.4.3
unstructured==0.7.12
pydantic==1.10.15
faiss-cpu==1.7.4